Remove REMOTE_ADDR from the data used to identify an user.
Using the IP address to identify a user is arguably a security "feature",
however it is very annoying to be randomly disconnected when your network
is connected to Internet via more than one external IP (for example in a
corporate environment).

Since the security added by cheking the IP is minimal when using HTTPS
(the php session id cookie is then transmitted encrypted), I propose to
remove it.